GERO-299 - refactor - move GraphKBConnection from __init__ to util.py.
